Combine Window and DesktopWindow
